@@ -1,44 +1,29 @@
 {
-  "name": "now-ui-dashboard-react",
-  "version": "1.5.1",
-  "main": "dist/index.js",
-  "private": true,
-  "homepage": "https://demos.creative-tim.com/now-ui-dashboard-react/#/",
-  "dependencies": {
-    "bootstrap": "4.6.0",
-    "chart.js": "3.2.1",
-    "firebase": "^9.17.1",
-    "gulp": "4.0.2",
-    "gulp-append-prepend": "1.0.9",
-    "perfect-scrollbar": "1.5.5",
-    "react": ">=16.8.0",
-    "react-bootstrap": "^2.7.2",
-    "react-chartjs-2": "3.0.3",
-    "react-dom": ">=16.8.0",
-    "react-google-button": "^0.7.2",
-    "react-notification-alert": "0.0.13",
-    "react-router-dom": "5.2.0",
-    "react-scripts": "5.0.1",
-    "reactstrap": "8.9.0",
-    "sass": "1.53.0"
-  },
+  "name": "@minimal/material-kit-react",
+  "author": "minimals.cc",
+  "licence": "MIT",
+  "version": "1.6.0",
+  "private": false,
   "scripts": {
     "start": "react-scripts start",
-    "build": "react-scripts build && gulp licenses",
+    "build": "react-scripts build",
     "test": "react-scripts test",
     "eject": "react-scripts eject",
-    "install:clean": "rm -rf node_modules/ && rm -rf package-lock.json && npm install && npm start",
-    "compile-sass": "sass src/assets/scss/paper-dashboard.scss src/assets/css/paper-dashboard.css",
-    "minify-sass": "sass src/assets/scss/paper-dashboard.scss src/assets/css/paper-dashboard.min.css --style compressed"
+    "lint": "eslint --ext .js,.jsx ./src",
+    "lint:fix": "eslint --fix --ext .js,.jsx ./src",
+    "clear-all": "rm -rf build node_modules",
+    "re-start": "rm -rf build node_modules && yarn install && yarn start",
+    "re-build": "rm -rf build node_modules && yarn install && yarn build"
   },
-  "optionalDependencies": {
-    "@babel/core": "7.18.9",
-    "jquery": "3.6.0",
-    "typescript": "4.7.4"
+  "eslintConfig": {
+    "extends": [
+      "react-app"
+    ]
   },
-  "peerDependencies": {
-    "react": ">=16.8.0",
-    "react-dom": ">=16.8.0"
+  "babel": {
+    "presets": [
+      "@babel/preset-react"
+    ]
   },
   "browserslist": {
     "production": [
@@ -51,5 +36,49 @@
       "last 1 firefox version",
       "last 1 safari version"
     ]
+  },
+  "dependencies": {
+    "@emotion/react": "^11.10.4",
+    "@emotion/styled": "^11.10.4",
+    "@faker-js/faker": "^7.6.0",
+    "@iconify/react": "^4.0.0",
+    "@mui/lab": "^5.0.0-alpha.103",
+    "@mui/material": "^5.10.9",
+    "apexcharts": "^3.36.0",
+    "change-case": "^4.1.2",
+    "date-fns": "^2.29.3",
+    "firebase": "^9.17.1",
+    "history": "^5.3.0",
+    "lodash": "^4.17.21",
+    "numeral": "^2.0.6",
+    "prop-types": "^15.8.1",
+    "react": "^18.2.0",
+    "react-apexcharts": "^1.4.0",
+    "react-dom": "^18.2.0",
+    "react-helmet-async": "^1.3.0",
+    "react-hook-form": "^7.37.0",
+    "react-router-dom": "^6.4.2",
+    "react-scripts": "^5.0.1",
+    "simplebar-react": "^2.4.3",
+    "web-vitals": "^3.0.3"
+  },
+  "devDependencies": {
+    "@babel/core": "^7.19.3",
+    "@babel/eslint-parser": "^7.19.1",
+    "@svgr/webpack": "^6.5.0",
+    "eslint": "^8.25.0",
+    "eslint-config-airbnb": "^19.0.4",
+    "eslint-config-prettier": "^8.5.0",
+    "eslint-config-react-app": "^7.0.1",
+    "eslint-plugin-flowtype": "^8.0.3",
+    "eslint-plugin-import": "^2.26.0",
+    "eslint-plugin-jsx-a11y": "^6.6.1",
+    "eslint-plugin-prettier": "^4.2.1",
+    "eslint-plugin-react": "^7.31.10",
+    "eslint-plugin-react-hooks": "^4.6.0",
+    "prettier": "^2.7.1"
+  },
+  "overrides": {
+    "@svgr/webpack": "^6.5.0"
   }
 }
