@@ -1,24 +1,32 @@
 import { createContext, useContext, useEffect, useState } from "react";
+import PropTypes from 'prop-types';
 import {
+  createUserWithEmailAndPassword,
+  signInWithEmailAndPassword,
   onAuthStateChanged,
   signOut,
   GoogleAuthProvider,
   signInWithPopup,
-  signInWithRedirect
 } from "firebase/auth";
 import { auth } from "../firebase";
 
 const userAuthContext = createContext();
 
-export function UserAuthContextProvider({ children }) {
+export function UserAuthContextProvider({ children = null }) {
   const [user, setUser] = useState({});
 
+  function logIn(email, password) {
+    return signInWithEmailAndPassword(auth, email, password);
+  }
+  function signUp(email, password) {
+    return createUserWithEmailAndPassword(auth, email, password);
+  }
   function logOut() {
     return signOut(auth);
   }
   function googleSignIn() {
     const googleAuthProvider = new GoogleAuthProvider();
-    return signInWithRedirect(auth, googleAuthProvider);
+    return signInWithPopup(auth, googleAuthProvider);
   }
 
   useEffect(() => {
@@ -34,12 +42,16 @@ export function UserAuthContextProvider({ children }) {
 
   return (
     <userAuthContext.Provider
-      value={{ user, logOut, googleSignIn }}
+      value={{ user, logIn, signUp, logOut, googleSignIn }}
     >
       {children}
     </userAuthContext.Provider>
   );
 }
+
+UserAuthContextProvider.propTypes = {
+  children: PropTypes.node,
+};
 
 export function useUserAuth() {
   return useContext(userAuthContext);
